post week-x changes (see description)
revised service cfn to not be dependent on specific task def version, to minimize cost, I deleted alb resource in cdrcluster cfn stack and to remove the service dependency on specific task def, I deleted it both. Make sure to read the notes for redeployment located at aws/cfn when the validation for this project is ready.